Commiting changes to advisor, javascript, css and the student files.
So as of right now the schedule appointment form pulls the options from
the database and the times select is dynamically chaining when you
select an advisor or date.
You can save an appointment to the database.
You can cancel an appointment from the view appointment screen.
I also changed the register form to match the login style .. I added the
register function to the student class. It is working now you can add a
user to the student table the only thing it needs is the hashing part
for the password. Jackie you should be able to just copy and past that
part.

For student, i am just hardcoding their email address on all the pages.. 

String email = "";
This needs to be changed to String email = User.getEmail()...